scrapy_project/requirements.txt
scrapy
google-cloud-storage